allow 'rsync without password'

+++ deploy.sh
@@ -23,7 +23,7 @@ do_sync()
     before_sync
     log " -> Done."
   fi
-  
+
   # ----------------
   # Sync
 
@@ -35,7 +35,7 @@ do_sync()
     wget -O .depignore https://raw.githubusercontent.com/KarappoInc/drone-deploy/master/.depignore
     DEP_IGNORE_FILE=$PWD/.depignore
   fi
-  
+
   log "- ignore file -> $DEP_IGNORE_FILE"
 
   if [ "$DEP_COMMAND" = "rsync" ]; then
@@ -48,11 +48,13 @@ do_sync()
       sudo apt-get install ssh-askpass
     fi
 
-    if [ `which sshpass` ]; then
-      log "- sshpass -> installed"
-    else
-      log "- sshpass -> install"
-      sudo apt-get install sshpass
+    if [ -f $DEP_PASSWORD ]; then
+      if [ `which sshpass` ]; then
+        log "- sshpass -> installed"
+      else
+        log "- sshpass -> install"
+        sudo apt-get install sshpass
+      fi
     fi
     # ------ /install -------
 
@@ -60,8 +62,16 @@ do_sync()
     if [ -f $DEP_IGNORE_FILE ]; then
       opt_exclude="--exclude-from=$DEP_IGNORE_FILE"
     fi
-    
-    if sshpass -p $DEP_PASSWORD rsync -aIzhv --stats --delete -e ssh $opt_exclude . $DEP_USER@$DEP_HOST:$DEP_HOST_DIR; then
+
+
+    RSYNC_COMMAND=rsync -aIzhv --stats --delete -e ssh $opt_exclude . $DEP_USER@$DEP_HOST:$DEP_HOST_DIR
+
+    # WITH password
+    if [ -f $DEP_PASSWORD ]; then
+      RSYNC_COMMAND=sshpass -p $DEP_PASSWORD $RSYNC_COMMAND
+    fi
+
+    if $RSYNC_COMMAND; then
       log "- sync -> done."
     else
       log "- sync -> [ERROR]"
@@ -107,7 +117,7 @@ do_sync()
       log "- sync -> [ERROR]"
       exit 1
     fi
-    
+
   fi
 
   # ----------------
